Move format and inspect to util/inspect
