+++ lib/ours/errors.js
@@ -1,6 +1,7 @@
 'use strict'
 
-const { format, inspect, AggregateError: CustomAggregateError } = require('./util')
+const { format, inspect } = require('./util/inspect')
+const { AggregateError: CustomAggregateError } = require('./util')
 
 /*
   This file is a reduced and adapted version of the main lib/internal/errors.js file defined at
